+++ openpecha/serializers/serialize.py
@@ -1,8 +1,8 @@
 import re
-from collections import defaultdict, namedtuple
+from collections import defaultdict
 from pathlib import Path
 
-from openpecha.formatters.layers import AnnType, SubText
+from openpecha.core.layer import LayerEnum
 from openpecha.utils import load_yaml
 
 INFO = "[INFO] {}"
@@ -174,12 +174,12 @@ class Serialize(object):
         return text for given span
         """
         if self.text_id:
-            vol_base = (self.opf_path / f"base/{base_id}.txt").read_text()
+            vol_base = (self.opf_path / f"base/{base_id}.txt").read_text(encoding='utf-8')
             start = self.text_spans[base_id]["start"]
             end = self.text_spans[base_id]["end"]
             return vol_base[start : end + 1]
         else:
-            vol_base = (self.opf_path / f"base/{base_id}.txt").read_text()
+            vol_base = (self.opf_path / f"base/{base_id}.txt").read_text(encoding='utf-8')
             return vol_base
 
     def get_text_base_layer(self):
@@ -226,14 +226,14 @@ class Serialize(object):
             sub_topics = topic["parts"]
             for sub_topic_uuid, sub_topic in sub_topics.items():
                 sub_topic_ann = defaultdict(str)
-                base_id = sub_topic['span'][0]['base']
-                sub_topic_ann["type"] = AnnType.sub_topic
+                base_id = sub_topic["span"][0]["base"]
+                sub_topic_ann["type"] = LayerEnum.sub_topic
                 sub_topic_ann["work_id"] = sub_topic["work_id"]
                 sub_topic_ann["span"] = sub_topic["span"][0]
                 self.apply_annotation(base_id, sub_topic_ann)
             if topic["span"]:
-                base_id = topic['span'][0]['base']
-                topic_ann["type"] = AnnType.topic
+                base_id = topic["span"][0]["base"]
+                topic_ann["type"] = LayerEnum.topic
                 topic_ann["span"] = topic["span"][0]
                 topic_ann["work_id"] = topic["work_id"]
                 self.apply_annotation(base_id, topic_ann)