+++ tests/test_tutorial/importers/test_tutorial001.py
@@ -6,6 +6,6 @@ from docs_src.importers.hfml.tutorial001 import result
 def test_hfml_base():
     output_fn = Path("tests") / "formatters" / "hfml" / "data" / "kangyur_base.txt"
 
-    expected = output_fn.read_text()
+    expected = output_fn.read_text(encoding='utf-8')
 
     assert result == expected